automation/requirements.txt
crewai>=0.134.0
python-dotenv>=1.0.1
requests>=2.32.3